SUPER kinda MASSIVE CHANGE (1.2)
so yeh I started work on this again out of boredom and irritation with
my lack of polish on this, so here we go with the change log...

-Actually fixed the scopes, I guess I thought the TTT scopes were
included with Gmod, but I guess not. You have to have people download
them (check content, I use scope_new)
-Removed the FAMAS for being generally shit
-I think I fixed the mapchange thing spazzing out and sending people a
billion windows (can someone confirm?)
-Made a new HUD that isn't all text, thanks to Cowabanga for the design
<3
-Made the Glock full auto
-Cut the M249's mag to 75 rounds (the 3 digit mag messed up the HUD
spacing)
-Players are now transparent while under spawn protection (1.5 seconds,
TODO: make it disable on fire/add visual queue for local player when
protected)
-Changed how ADS works
~Reloading boots you out of aiming (ooh)
~Now you have to hold down RMB to aim (no way!)
~You get a slight recoil decrease/accuracy increase for aiming (#whoa)
~You can't sprint while aiming (so much innovation!)
~Ironsights don't lag anymore like they would on a crappy DarkRP server
(actually kinda cool)

And that's basically it. I'll be pumping out more patches soon, so heads
up!!!!!