1.4 adding optional custom content
This update has a bunch of new shit, in addition to a new recoil system,
it also features battle chatter and a bunch of new custom stuff that's
from my server that you can add if you want. Here's the full list:

*New recoil system that tries and fails to emulate CS, in which spraying
will end up kicking your bullets over your crosshair. Doesn't work very
well, since it only bases the spray on how much view punch you have at
the time, so while the first shot is accurate, all the following shots
on an automatic weapon have the exact same spray levels. I'll work on
making a better system soon, this works for now
*Added battle chatter, which is basically the "killed him!" and "oh god
I'm dead" stuff you hear in Call of Duty for example. It's a small
change, but I think it's cool.

CUSTOM CONTENT:
*Since very few people seem to actually be making any custom content for
their servers, I've decided to release the stuff I've made for my own
server, namely a dildo launcher, incendiary shotgun, CZ-75 or something,
PPSH, revolver and grenade launcher that shouts "ONII CHAN~" when you
fire it, as well as some additional sounds and shit. This stuff requires
some extra downloads, so it's not included in the base gamemode thingy,
check the _extracustomcontent for the models/materials/sounds/gameplay
files you need in order to use that new shit. And if I'm doing a
horrible job of explaining this, just add me on Steam, my name is Shaps.